requirements.txt
Django==4.2.6
django-cleanup==8.0.0
django-cors-headers==4.3.0
django-environ==0.10.0
django-import-export==3.1.0
django-rest-framework==0.1.0
django-split-settings==1.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-camel-case2==0.3.0
djangorestframework-simplejwt==5.3.0
djangorestframework-filters==0.11.1
Markdown==3.5
mkdocs==1.5.3
mkdocs-dracula-theme==1.0.5
mkdocs-material==9.4.7
mkdocs-material-extensions==1.3
openpyxl==3.1.2
paginate==0.5.6
Pillow==10.0.1
psycopg2-binary==2.9.6
PyJWT==2.8.0
python-dotenv==1.0.0
redis==4.5.1
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
